Migrated v3 of the GitHub API from v2.
v3 of the GitHub API supports CORS which obviates the need for the Node.JS proxy
server in well behaved, modern browsers. Safari 5.1.5 is not one of these browsers.
These changes assume that CORS works correctly and does not fall back to the Node.JS
proxy.

Authentication was also changed to use the authorizations API instead of standard
OAuth authentication.